Fully cover internal packages
